requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b1.dev132+g2c5041c
click==7.1.2
dataclasses==0.7
mypy-extensions==0.4.3
pathspec==0.8.0
Pillow==7.2.0
pkg-resources==0.0.0
regex==2020.7.14
Rx==3.1.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
